@@ -1,47 +1,42 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.4.2")
-    }
-}
 plugins {
-	id 'java'
-    id 'net.ltgt.apt' version '0.8'
+    id 'org.springframework.boot' version '3.0.2'
+    id 'io.spring.dependency-management' version '1.1.0'
+    id 'java'
     id 'idea'
 }
-apply plugin: "net.ltgt.apt"
-apply plugin: 'java'
-apply plugin: 'eclipse'
-apply plugin: 'idea'
-apply plugin: 'org.springframework.boot'
-apply plugin: 'io.spring.dependency-management'
 
-bootJar {
-    baseName = 'niivo-stock-data'
-    version =  '1.0'
-}
+group = 'dev.anshumax'
+version = '1.0'
 
 repositories {
     mavenCentral()
+    mavenLocal()
     flatDir {
         dirs 'libs'
     }
+    maven {
+        url "https://oss.sonatype.org/content/repositories/snapshots"
+    }
+    maven {
+        url "https://nexus.geomatys.com/repository/geotoolkit/"
+    }
+
+    maven { url 'https://jitpack.io' }
 }
 
-sourceCompatibility = 11
-targetCompatibility = 11
+sourceCompatibility = 17
+targetCompatibility = 17
 
 dependencies {
-    compile group: 'org.springframework.boot', name:'spring-boot-starter-data-jpa'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: '3.0.4'
 
-    compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.10'
-    compile group: 'commons-io', name: 'commons-io', version: '2.7'
-    
-    compile group: 'mysql', name: 'mysql-connector-java'
-	compile group: 'com.google.code.gson', name: 'gson', version: '2.8.5'
-	compile group: 'org.seleniumhq.selenium', name: 'selenium-java', version: '3.141.59'
-    
+    implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.12.0'
+    implementation group: 'commons-io', name: 'commons-io', version: '2.11.0'
+    implementation group: 'org.projectlombok', name: 'lombok', version: '1.18.26'
+
+
+    implementation group: 'mysql', name: 'mysql-connector-java', version: '8.0.32'
+    implementation group: 'com.google.code.gson', name: 'gson', version: '2.10.1'
+    implementation group: 'org.seleniumhq.selenium', name: 'selenium-java', version: '4.8.1'
     
 }
