@@ -2,5 +2,5 @@ package com.claroinvestments.stockprice;
 
 public enum ExecutionType {
 	DAILY,
-	HISTORICAL;
+	HISTORICAL
 }
